Allow existing Queues to create topic
If you provide the Queue name (which already exists in SQS), but you don't have a Topic to respond to this queue, it will call the `create()` function, which will attempt to create the query again (although it already exists). Need to check that in the create function as well (and not only in the publish method).

I know it does not makes sense to call the `create` method and check if the Queue exists or not, but if not, a refactor would be need to separete que creation of the queue and the topic one, so that on publish method, if the topic is not created, we just called the createTopic method and not the create function.

Does this make sense to any of you guys?